Add a default .plot() method to all SysTest objects.
The SysTest base class now includes a plot() method that takes the results of a SysTest and passes them through (if they're plottable) or returns an object with a .savefig() method that doesn't do anything (if they're not) so we never have to do checks in the main program for whether the results are plottable. (#7).